requirements1.txt
langchain==0.0.154
PyPDF==6.1.3
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.12.0
streamlit-extras
